pyproject.toml
[project]
name = "data-extractor"
version = "0.1.0"
description = "GenAI data extraction project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-generativeai>=0.8.5",
    "openai>=1.86.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "selenium>=4.33.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
